Remove javax.inject dependency
As far as I can see, Spring 3.0 supports only parts of the Java EE 6
Contexts and Dependency Injection (CDI). The javax.inject (JSR-330) is
supported, the javax.enterprise.context/inject (JSR-299) is not.

This means that, for example, when you want to introduce a factory
method, you may not use @Produces. Instead, you have to use Spring's
own @Bean. I expect that more of these kind of things will get to the
surface and prefer to use 'native' Spring instead of some Spring/CDI
mix.