Add JDBC account DAO and unit tests
Both the web application and the unit tests are using an in-memory
embedded database (HSQLDB). See schema.sql and data.sql for the scripts
that are used to create the database schema and to fill the database
with test data.

For the unit tests, Hamcrest is used to make the tests a pleasure to
read. Because I wanted to use a newer version of Hamcrest then bundled
with JUnit and Mockito, the pom.xml will exclude Hamcrest from these
frameworks. A newer version was needed to be able to use the
IsCollectionWithSize matcher and its hasSize feature.