Change JSF outcome strings to use PrettyFaces ID's
The PrettyFaces ID's (like: "pretty:accountNotFound") will be used for
all the outcome strings. This saves us from maintaining the navigation
rules of faces-config.xml.

There is a disadvantage for not having your navigation rules on one
location in your faces-config.xml: there is no longer an easy way to get
an overview of the page flow. But because we already had GET requests
(by using <h:link>) from the JSF pages, the page flow was already not
specified on one location.