Add form-based authentication
The implementation is not finished yet. An in-memory user service that
contains the users is used (there is currently one: testuser/testuser).
The passwords are not encrypted. And there is no logout button or link
(although you may use the /logout URL to logout).

Note: Spring Security 3.0.6 contains a bug. Because of this bug you
could not login any more after a failed login (the password of the
(cached) user details was set to null after the failed login).
Therefore, version 3.0.7 is used (the other Spring modules are still on
3.0.6).